requirements.txt
Flask
python-dotenv
google-generativeai
gunicorn
feedparser
scikit-learn
rich
